refactor(FRENCH): split the French mixin by concern, mirroring upbrr's tracker modules
FRENCH.py was 2378 lines of mixed responsibilities. It is now a thin
composition of six modules under src/trackers/french/, each mapping onto
one module of a future upbrr tracker:

  languages   (pure tables + VFF/VFQ/VF2/VOSTFR detection) -> languageutil/french.go
  naming                                                   -> name.go
  rules       (get_additional_checks)                      -> rules.go / validation.go
  dupes       (French language hierarchy)                  -> dupe.go
  description (BBCode blocks from MediaInfo)               -> description.go
  nfo         (NFO discovery, torrent re-creation)         -> media.go

Method bodies are moved verbatim; FrenchTrackerMixin keeps its public
surface and the module constants are re-exported, so no tracker or test
changes.
