+++ src/trackers/FRENCH.py
@@ -2,2377 +2,52 @@
 """
 French tracker mixin — shared logic for all French-language trackers.
 
-All French trackers (C411, G3MINI, V3X, …) inherit from this mixin
-to share a single, canonical implementation of:
-  · Audio language detection / French dub suffix (VFF, VFQ, VF2, …)
-  · Language tag building  (MULTI.VFF, VOSTFR, MUET, …)
-  · French title from TMDB
-  · Release naming (dot-separated, French-tracker conventions)
+All French trackers (C411, G3MINI, V3X, …) inherit from this mixin. The
+implementation is split by concern under ``src/trackers/french/`` so that
+each piece maps onto one module of a future upbrr tracker:
+
+  · languages   — audio/subtitle facts, VFF/VFQ/VF2/VOSTFR tags (pure)
+  · naming      — release name conventions
+  · rules       — release eligibility
+  · dupes       — French language hierarchy on duplicate candidates
+  · description — BBCode blocks from MediaInfo
+  · nfo         — NFO discovery/generation, torrent re-creation
 """
 
-import asyncio
-import glob
-import hashlib
-import os
-import re
-from datetime import datetime
-from typing import Any, Optional, Union
+from typing import Any
 
-import aiofiles
-from unidecode import unidecode
+from src.trackers.french.description import FrenchDescriptionMixin
+from src.trackers.french.dupes import FrenchDupeMixin
+from src.trackers.french.languages import (
+    _FRENCH_AUDIO_THRESHOLD,
+    FRENCH_LANG_HIERARCHY,
+    FRENCH_LANG_VALUES,
+    LANG_FLAGS,
+    LANG_MAP,
+    LANG_NAMES_FR,
+    FrenchLanguageMixin,
+)
+from src.trackers.french.naming import FrenchNamingMixin
+from src.trackers.french.nfo import FrenchNfoMixin
+from src.trackers.french.rules import FrenchRulesMixin
 
-from src.audio import AD_TRACK_RE, codec_info_from_track
-from src.console import console
-from src.nfo_generator import SceneNfoGenerator, is_multi_episode_nfo
-from src.torrentcreate import TorrentCreator
-from src.trackers.COMMON import COMMON
+__all__ = [
+    "_FRENCH_AUDIO_THRESHOLD",
+    "FRENCH_LANG_HIERARCHY",
+    "FRENCH_LANG_VALUES",
+    "LANG_FLAGS",
+    "LANG_MAP",
+    "LANG_NAMES_FR",
+    "FrenchTrackerMixin",
+]
 
 Meta = dict[str, Any]
 
-# ── Language → 3-letter ISO 639 mapping (comprehensive) ──────
-LANG_MAP: dict[str, str] = {
-    # French
-    "fr": "FRA",
-    "fre": "FRA",
-    "fra": "FRA",
-    "french": "FRA",
-    "français": "FRA",
-    "francais": "FRA",
-    "fr-fr": "FRA",
-    "fr-ca": "FRA",
-    "fr-be": "FRA",
-    "fr-ch": "FRA",
-    # English
-    "en": "ENG",
-    "eng": "ENG",
-    "english": "ENG",
-    "en-us": "ENG",
-    "en-gb": "ENG",
-    # Spanish
-    "es": "SPA",
-    "spa": "SPA",
-    "spanish": "SPA",
-    "español": "SPA",
-    "castellano": "SPA",
-    "es-es": "SPA",
-    "lat": "LAT",
-    "latino": "LAT",
-    "latin american spanish": "LAT",
-    "es-mx": "LAT",
-    "es-419": "LAT",
-    # German
-    "de": "DEU",
-    "deu": "DEU",
-    "ger": "DEU",
-    "german": "DEU",
-    "deutsch": "DEU",
-    # Italian
-    "it": "ITA",
-    "ita": "ITA",
-    "italian": "ITA",
-    "italiano": "ITA",
-    # Portuguese
-    "pt": "POR",
-    "por": "POR",
-    "portuguese": "POR",
-    "português": "POR",
-    "portuguese (iberian)": "POR",
-    "pt-br": "POR",
-    "pt-pt": "POR",
-    # Japanese
-    "ja": "JPN",
-    "jpn": "JPN",
-    "japanese": "JPN",
-    "日本語": "JPN",
-    # Korean
-    "ko": "KOR",
-    "kor": "KOR",
-    "korean": "KOR",
-    "한국어": "KOR",
-    # Chinese
-    "zh": "ZHO",
-    "zho": "ZHO",
-    "chi": "ZHO",
-    "chinese": "ZHO",
-    "mandarin": "ZHO",
-    "中文": "ZHO",
-    "zh-cn": "ZHO",
-    # Russian
-    "ru": "RUS",
-    "rus": "RUS",
-    "russian": "RUS",
-    "русский": "RUS",
-    # Arabic
-    "ar": "ARA",
-    "ara": "ARA",
-    "arabic": "ARA",
-    # Hindi
-    "hi": "HIN",
-    "hin": "HIN",
-    "hindi": "HIN",
-    # Dutch
-    "nl": "NLD",
-    "nld": "NLD",
-    "dut": "NLD",
-    "dutch": "NLD",
-    # Polish
-    "pl": "POL",
-    "pol": "POL",
-    "polish": "POL",
-    # Turkish
-    "tr": "TUR",
-    "tur": "TUR",
-    "turkish": "TUR",
-    # Thai
-    "th": "THA",
-    "tha": "THA",
-    "thai": "THA",
-    # Vietnamese
-    "vi": "VIE",
-    "vie": "VIE",
-    "vietnamese": "VIE",
-    # Swedish
-    "sv": "SWE",
-    "swe": "SWE",
-    "swedish": "SWE",
-    # Norwegian
-    "no": "NOR",
-    "nor": "NOR",
-    "norwegian": "NOR",
-    "nb": "NOR",
-    "nob": "NOR",
-    # Danish
-    "da": "DAN",
-    "dan": "DAN",
-    "danish": "DAN",
-    # Finnish
-    "fi": "FIN",
-    "fin": "FIN",
-    "finnish": "FIN",
-    # Czech
-    "cs": "CES",
-    "ces": "CES",
-    "cze": "CES",
-    "czech": "CES",
-    # Hungarian
-    "hu": "HUN",
-    "hun": "HUN",
-    "hungarian": "HUN",
-    # Romanian
-    "ro": "RON",
-    "ron": "RON",
-    "rum": "RON",
-    "romanian": "RON",
-    # Greek
-    "el": "ELL",
-    "ell": "ELL",
-    "gre": "ELL",
-    "greek": "ELL",
-    # Hebrew
-    "he": "HEB",
-    "heb": "HEB",
-    "hebrew": "HEB",
-    # Indonesian
-    "id": "IND",
-    "ind": "IND",
-    "indonesian": "IND",
-    # Ukrainian
-    "uk": "UKR",
-    "ukr": "UKR",
-    "ukrainian": "UKR",
-    # Tamil / Telugu
-    "ta": "TAM",
-    "tam": "TAM",
-    "tamil": "TAM",
-    "te": "TEL",
-    "tel": "TEL",
-    "telugu": "TEL",
-    # Malay
-    "ms": "MSA",
-    "msa": "MSA",
-    "may": "MSA",
-    "malay": "MSA",
-    # Persian
-    "fa": "FAS",
-    "fas": "FAS",
-    "per": "FAS",
-    "persian": "FAS",
-}
 
-# ── Language → flag emoji mapping (for BBCode descriptions) ──
-LANG_FLAGS: dict[str, str] = {
-    "english": "🇺🇸",
-    "french": "🇫🇷",
-    "german": "🇩🇪",
-    "spanish": "🇪🇸",
-    "italian": "🇮🇹",
-    "portuguese": "🇵🇹",
-    "russian": "🇷🇺",
-    "japanese": "🇯🇵",
-    "korean": "🇰🇷",
-    "chinese": "🇨🇳",
-    "arabic": "🇸🇦",
-    "dutch": "🇳🇱",
-    "polish": "🇵🇱",
-    "turkish": "🇹🇷",
-    "thai": "🇹🇭",
-    "swedish": "🇸🇪",
-    "norwegian": "🇳🇴",
-    "norwegian bokmal": "🇳🇴",
-    "norwegian bokmål": "🇳🇴",
-    "norwegian nynorsk": "🇳🇴",
-    "danish": "🇩🇰",
-    "finnish": "🇫🇮",
-    "czech": "🇨🇿",
-    "hungarian": "🇭🇺",
-    "romanian": "🇷🇴",
-    "greek": "🇬🇷",
-    "hebrew": "🇮🇱",
-    "indonesian": "🇮🇩",
-    "bulgarian": "🇧🇬",
-    "croatian": "🇭🇷",
-    "serbian": "🇷🇸",
-    "slovenian": "🇸🇮",
-    "estonian": "🇪🇪",
-    "icelandic": "🇮🇸",
-    "lithuanian": "🇱🇹",
-    "latvian": "🇱🇻",
-    "ukrainian": "🇺🇦",
-    "hindi": "🇮🇳",
-    "tamil": "🇮🇳",
-    "telugu": "🇮🇳",
-    "malay": "🇲🇾",
-    "vietnamese": "🇻🇳",
-    "persian": "🇮🇷",
-    "cantonese": "🇨🇳",
-    "mandarin": "🇨🇳",
-    "slovak": "🇸🇰",
-    "catalan": "🇪🇸",
-    "basque": "🇪🇸",
-    "galician": "🇪🇸",
-    "bengali": "🇧🇩",
-    "urdu": "🇵🇰",
-    "tagalog": "🇵🇭",
-    "filipino": "🇵🇭",
-    "khmer": "🇰🇭",
-    "mongolian": "🇲🇳",
-    "georgian": "🇬🇪",
-    "albanian": "🇦🇱",
-    "macedonian": "🇲🇰",
-    "bosnian": "🇧🇦",
-    "swahili": "🇰🇪",
-}
-
-# ── Language → French display name ───────────────────────────
-LANG_NAMES_FR: dict[str, str] = {
-    "english": "Anglais",
-    "french": "Français",
-    "german": "Allemand",
-    "spanish": "Espagnol",
-    "italian": "Italien",
-    "portuguese": "Portugais",
-    "russian": "Russe",
-    "japanese": "Japonais",
-    "korean": "Coréen",
-    "chinese": "Chinois",
-    "arabic": "Arabe",
-    "dutch": "Néerlandais",
-    "polish": "Polonais",
-    "turkish": "Turc",
-    "thai": "Thaï",
-    "swedish": "Suédois",
-    "norwegian": "Norvégien",
-    "norwegian bokmal": "Norvégien",
-    "norwegian bokmål": "Norvégien",
-    "norwegian nynorsk": "Norvégien (nynorsk)",
-    "danish": "Danois",
-    "finnish": "Finnois",
-    "czech": "Tchèque",
-    "hungarian": "Hongrois",
-    "romanian": "Roumain",
-    "greek": "Grec",
-    "hebrew": "Hébreu",
-    "indonesian": "Indonésien",
-    "bulgarian": "Bulgare",
-    "croatian": "Croate",
-    "serbian": "Serbe",
-    "slovenian": "Slovène",
-    "estonian": "Estonien",
-    "icelandic": "Islandais",
-    "lithuanian": "Lituanien",
-    "latvian": "Letton",
-    "ukrainian": "Ukrainien",
-    "hindi": "Hindi",
-    "tamil": "Tamoul",
-    "telugu": "Télougou",
-    "malay": "Malais",
-    "vietnamese": "Vietnamien",
-    "persian": "Persan",
-    "cantonese": "Cantonais",
-    "mandarin": "Mandarin",
-    "slovak": "Slovaque",
-    "catalan": "Catalan",
-    "basque": "Basque",
-    "galician": "Galicien",
-    "bengali": "Bengali",
-    "urdu": "Ourdou",
-    "tagalog": "Tagalog",
-    "filipino": "Filipino",
-    "khmer": "Khmer",
-    "mongolian": "Mongol",
-    "georgian": "Géorgien",
-    "albanian": "Albanais",
-    "macedonian": "Macédonien",
-    "bosnian": "Bosniaque",
-    "swahili": "Swahili",
-}
-
-# Canonical list of French language values (for subtitle/audio detection)
-FRENCH_LANG_VALUES = frozenset(
-    {
-        "french",
-        "fre",
-        "fra",
-        "fr",
-        "français",
-        "francais",
-        "fr-fr",
-        "fr-ca",
-        "fr-be",
-        "fr-ch",
-    }
-)
-
-# ── French language hierarchy for dupe checking ──────────────
-# On French trackers a release with French audio always supersedes a
-# VOSTFR (subtitles-only) or VO (original-only) version of the same
-# content.  The hierarchy ranks tags from most desirable (MULTI, 7)
-# to least (VO, 1).
-FRENCH_LANG_HIERARCHY: dict[str, int] = {
-    "MULTI": 7,
-    "VFF": 6,
-    "VFQ": 6,
-    "VFB": 6,
-    "VF2": 6,
-    "VOF": 5,
-    "VOQ": 5,  # original Québécois audio — Québec counterpart of VOF
-    "TRUEFRENCH": 4,
-    "FRENCH": 3,
-    "VOSTFR": 2,
-    "SUBFRENCH": 2,  # legacy alias for VOSTFR
-    "VO": 1,
-}
-
-# Threshold: tags at or above this level indicate French audio is present
-_FRENCH_AUDIO_THRESHOLD = 3  # FRENCH and above
-
-
-class FrenchTrackerMixin:
+class FrenchTrackerMixin(FrenchRulesMixin, FrenchLanguageMixin, FrenchNamingMixin, FrenchDupeMixin, FrenchDescriptionMixin, FrenchNfoMixin):
     """Mixin providing French-tracker naming and audio analysis.
 
     Mix this into any tracker class that targets a French tracker.
     Requires the host class to have a ``tmdb_manager`` attribute
     (instance of :class:`src.tmdb.TmdbManager`).
     """
-
-    # Signals that this tracker auto-detects NFO files on disk and includes them in
-    # the torrent.  Used by upload.py to proactively set keep_nfo before BASE creation
-    # so that BASE.torrent already contains the NFO (avoids a later full rehash).
-    auto_nfo: bool = True
-
-    # Subclasses may override to change the WEBDL source label in release names
-    # e.g. "WEB" (C411/V3X) vs "WEB-DL" (G3MINI)
-    WEB_LABEL: str = "WEB"
-
-    # Whether to include the streaming service name (NF, AMZN, …) in the release name.
-    # Set to False for trackers that want the service only in the description.
-    INCLUDE_SERVICE_IN_NAME: bool = True
-
-    # Whether to prefer the original-language title in release names.
-    # When True and the movie is not originally French, the English/original
-    # title is used instead of the French TMDB translation.
-    # Set to True for trackers that accept both title languages (e.g. NST).
-    PREFER_ORIGINAL_TITLE: bool = False
-
-    # Whether the "UHD" tag should only appear for REMUX / DISC releases.
-    # C411 wiki: "UHD is only allowed when the title contains REMUX/BDMV/ISO".
-    # When True, UHD is stripped from ENCODE, WEBDL, WEBRIP, HDTV, DVDRIP.
-    UHD_ONLY_FOR_REMUX_DISC: bool = False
-
-    # Subclasses may set this to a non-empty string to accept notag releases
-    # with a replacement label (e.g. "NOTAG", "NoGrp").
-    notag_label: str = ""
-
-    # ──────────────────────────────────────────────────────────
-    #  Language checks
-    # ──────────────────────────────────────────────────────────
-
-    async def get_additional_checks(self, meta: dict[str, Any]) -> bool:
-        """Default French language check for all French trackers.
-
-        Subclasses that inherit UNIT3D get this called automatically from
-        ``UNIT3D.search_existing()``.  Standalone French trackers (C411,
-        V3X) must call it explicitly from their own
-        ``search_existing()``.
-
-        Subclasses may override to add extra rules (banned types, etc.).
-        """
-        french_languages = list(FRENCH_LANG_VALUES)
-        if not await self.common.check_language_requirements(
-            meta,
-            self.tracker,
-            languages_to_check=french_languages,
-            check_audio=True,
-            check_subtitle=True,
-            require_both=False,
-        ):
-            if not meta.get("unattended", False):
-                console.print(f"[bold red]Language requirements not met for {self.tracker}.[/bold red]")
-            return False
-        return True
-
-    # ──────────────────────────────────────────────────────────
-    #  Edition formatting
-    # ──────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _format_edition(edition: str) -> str:
-        """Convert an uppercased edition to title case for French trackers.
-
-        French trackers use title case for edition keywords:
-        ``SPECIAL EDITION`` → ``Special Edition``,
-        ``DIRECTOR'S CUT`` → ``Director's Cut``, etc.
-
-        Mixed-case scene-style tags (e.g. ``LiMiTED``) are preserved as-is.
-        """
-        if not edition:
-            return edition
-        # Only title-case fully uppercased strings; preserve scene-style
-        # mixed-case tags like "LiMiTED" unchanged.
-        if edition != edition.upper():
-            return edition
-        result = edition.title()
-        # Fix capitalization after apostrophes: "Director'S" → "Director's"
-        result = re.sub(r"(\w)'(\w)", lambda m: f"{m.group(1)}'{m.group(2).lower()}", result)
-        return result
-
-    # ──────────────────────────────────────────────────────────
-    #  Audio-track helpers
-    # ──────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _get_audio_tracks(meta: Meta, filter_commentary: bool = True) -> list[dict[str, Any]]:
-        """Extract audio tracks from MediaInfo, optionally filtering commentary."""
-        if "mediainfo" not in meta or "media" not in meta.get("mediainfo", {}):
-            return []
-        tracks = meta["mediainfo"]["media"].get("track", [])
-        audio = [t for t in tracks if t.get("@type") == "Audio"]
-        if filter_commentary:
-            audio = [t for t in audio if "commentary" not in str(t.get("Title", "")).lower() and "comment" not in str(t.get("Title", "")).lower()]
-        return audio
-
-    def _should_include_ad_prefix(self, has_french_audio: bool, ad_audio_langs: list[str]) -> bool:
-        """Whether to include the ``AD.`` prefix in the release name.
-
-        Subclasses may override for tracker-specific rules.
-        """
-        return True
-
-    @staticmethod
-    def _is_audio_desc_track(track: dict[str, Any]) -> bool:
-        """Return True when an audio track is an audio-description track."""
-        title = str(track.get("Title") or track.get("title") or "")
-        return bool(AD_TRACK_RE.search(title))
-
-    def _get_audio_for_name(self, meta: Meta) -> str:
-        """Return the audio codec+channels string for the release name.
-
-        Base implementation uses ``meta['audio']`` (first track in stream
-        order).  Subclasses may override to pick a different track, e.g.
-        the first French audio track for French-tracker NFO validation.
-        """
-
-        lossless_additional_features = ["XLL", "HD MA", ":X", "16-ch", "MLP FBA"]
-        lossless_tracks = []
-        lossy_tracks = []
-        audio_tracks = self._get_audio_tracks(meta)
-
-        main_tracks = [
-            t
-            for t in audio_tracks
-            if not self._is_audio_desc_track(t) and "compatibility" not in str(t.get("Title", t.get("title", ""))).lower() and t.get("Channels") and t.get("Format")
-        ]
-
-        if not main_tracks:  # Fallback if no "main tracks" was found
-            return meta.get("audio", "").replace("Dual-Audio", "").replace("Dubbed", "").replace("DD+", "DDP")
-
-        def most_channels_priority(t):
-            channels = int(t.get("Channels", "0"))
-            is_french = 1 if self._map_language(str(t.get("Language", ""))) == "FRA" else 0
-            return (channels, is_french)
-
-        for t in main_tracks:
-            is_lossless = (
-                t.get("Compression_Mode") == "Lossless"
-                or any(f in str(t.get("Format_AdditionalFeatures", "")) for f in lossless_additional_features)
-                or any(f in str(t.get("Format_Commercial_IfAny", "")) for f in lossless_additional_features)
-            )
-            if is_lossless:
-                lossless_tracks.append(t)
-            else:
-                lossy_tracks.append(t)
-
-        if lossless_tracks:
-            return codec_info_from_track(max(lossless_tracks, key=most_channels_priority)).replace("Dual-Audio", "").replace("Dubbed", "").replace("DD+", "DDP")
-        elif lossy_tracks:
-            return codec_info_from_track(max(lossy_tracks, key=most_channels_priority)).replace("Dual-Audio", "").replace("Dubbed", "").replace("DD+", "DDP")
-
-        return meta.get("audio", "").replace("Dual-Audio", "").replace("Dubbed", "").replace("DD+", "DDP")
-
-    @staticmethod
-    def _map_language(lang: str) -> str:
-        """Map a language name/code to a normalised 3-letter code."""
-        if not lang:
-            return ""
-        mapped = LANG_MAP.get(str(lang).lower().strip())
-        if mapped:
-            return mapped
-        raw = str(lang).strip()
-        return raw.upper()[:3] if len(raw) >= 3 else raw.upper()
-
-    @staticmethod
-    def _extract_audio_languages(audio_tracks: list[dict[str, Any]], meta: Optional[Meta] = None) -> list[str]:
-        """Extract and normalise audio language codes (de-duplicated, ordered)."""
-        langs: list[str] = []
-        for track in audio_tracks:
-            raw = str(track.get("Language", "")).strip().lower()
-            if not raw:
-                # Fallback: infer from track Title
-                title = str(track.get("Title", "")).strip().lower()
-                if any(k in title for k in ("french", "français", "francais")):
-                    raw = "french"
-                elif any(k in title for k in ("english", "anglais")):
-                    raw = "english"
-            mapped = LANG_MAP.get(raw, raw.upper()[:3] if raw else "")
-            if mapped and mapped not in langs:
-                langs.append(mapped)
-        # Fallback: meta['audio_languages']
-        if not langs and meta and meta.get("audio_languages"):
-            for lang in meta["audio_languages"]:
-                code = LANG_MAP.get(str(lang).lower().strip(), str(lang).upper()[:3])
-                if code and code not in langs:
-                    langs.append(code)
-        return langs
-
-    @staticmethod
-    def _get_french_dub_suffix(audio_tracks: list[dict[str, Any]]) -> Union[str, None]:
-        """Determine French dub variant from audio-track Language/Title fields.
-
-        Checks the *raw* Language tag for regional variants (``fr-fr`` → VFF,
-        ``fr-ca`` → VFQ) and the track Title for explicit VFF/VFQ/VF2 labels.
-
-        Returns ``'VFF'``, ``'VFQ'``, ``'VF2'``, ``'VF<n>'`` (n>2), or ``None``.
-        """
-        fr_variants: list[str] = []
-
-        for track in audio_tracks:
-            lang = track.get("Language", "")
-            if not isinstance(lang, str):
-                continue
-            ll = lang.lower().strip()
-
-            # Check raw Language tag for region codes
-            if ll == "fr-fr" and "fr-fr" not in fr_variants:
-                fr_variants.append("fr-fr")
-            elif ll in ("fr-ca", "fr-qc") and "fr-ca" not in fr_variants:
-                fr_variants.append("fr-ca")
-            elif ll == "fr-be" and "fr-be" not in fr_variants:
-                fr_variants.append("fr-be")
-            elif ll == "fr-ch":
-                if "fr-fr" not in fr_variants:
-                    fr_variants.append("fr-fr")  # Swiss French → treat as VFF
-            elif ll in ("fr", "fre", "fra", "french", "français", "francais"):
-                # Generic French — check Title for explicit VFF/VFQ/VFB or region keywords
-                title = str(track.get("Title", "")).upper()
-                # Canadian French indicators
-                is_canadian = (
-                    "VFQ" in title
-                    or "CANADA" in title
-                    or "CANADIEN" in title
-                    or "QUÉB" in title
-                    or "QUEB" in title
-                    or "(CA)" in title
-                    or re.search(r"\bCA\b", title)  # "FR CA 5.1" → matches CA as word
-                )
-                # Belgian French indicators
-                is_belgian = "VFB" in title or "BELGE" in title or "BELGIQUE" in title or "(BE)" in title
-                if is_canadian:
-                    if "fr-ca" not in fr_variants:
-                        fr_variants.append("fr-ca")
-                elif is_belgian:
-                    if "fr-be" not in fr_variants:
-                        fr_variants.append("fr-be")
-                elif "VFF" in title or "(FR)" in title or "FRANCE" in title:
-                    if "fr-fr" not in fr_variants:
-                        fr_variants.append("fr-fr")
-                elif "VF2" in title:
-                    return "VF2"  # explicit VF2 in title
-                else:
-                    if "fr" not in fr_variants:
-                        fr_variants.append("fr")
-
-        n = len(fr_variants)
-        if n == 0:
-            return None
-        if n >= 2:
-            return f"VF{n}"
-
-        has_vfq = "fr-ca" in fr_variants
-        has_vfb = "fr-be" in fr_variants
-        has_vff = "fr-fr" in fr_variants
-
-        if has_vfq:
-            return "VFQ"
-        if has_vfb:
-            return "VFB"
-        if has_vff:
-            return "VFF"
-        return None  # generic 'fr' only — no suffix
-
-    @staticmethod
-    def _has_french_subs(meta: Meta) -> bool:
-        """Check whether French subtitles are present in MediaInfo."""
-        if "mediainfo" not in meta or "media" not in meta.get("mediainfo", {}):
-            return False
-        for track in meta["mediainfo"]["media"].get("track", []):
-            if track.get("@type") != "Text":
-                continue
-            lang = str(track.get("Language", "")).lower().strip()
-            if lang in FRENCH_LANG_VALUES or lang.startswith("fr"):
-                return True
-            title = str(track.get("Title", "")).lower()
-            if "french" in title or "français" in title or "francais" in title:
-                return True
-        return False
-
-    # ──────────────────────────────────────────────────────────
-    #  French language hierarchy — dupe checking
-    # ──────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _extract_french_lang_tag(name: str) -> tuple[str, int]:
-        """Extract the highest-level French language tag from a release name.
-
-        Returns ``(tag, level)`` where *level* comes from
-        :data:`FRENCH_LANG_HIERARCHY`.  ``('', 0)`` when no tag is found.
-
-        The match is case-insensitive and requires the tag to be delimited
-        by dots, spaces, hyphens, underscores, or string boundaries so that
-        ``VO`` does not match inside ``VOSTFR`` and ``FRENCH`` does not
-        match inside ``TRUEFRENCH``.
-        """
-        name_upper = name.upper()
-        best_tag = ""
-        best_level = 0
-        for tag, level in FRENCH_LANG_HIERARCHY.items():
-            pattern = rf"(?:^|[\.\s\-_])({re.escape(tag)})(?:[\.\s\-_]|$)"
-            if re.search(pattern, name_upper) and level > best_level:
-                best_tag = tag
-                best_level = level
-        return (best_tag, best_level)
-
-    async def _check_french_lang_dupes(
-        self,
-        dupes: list[dict[str, Any]],
-        meta: Meta,
-    ) -> list[dict[str, Any]]:
-        """Filter and flag dupes based on French language hierarchy.
-
-        On French trackers:
-
-        1. **Upload has French audio** (MULTI, VFF, …): existing releases
-           that *lack* French audio (VOSTFR, VO) are **removed** from the
-           dupe list — they are inferior and do not block the upload.
-
-        2. **Upload lacks French audio** (VOSTFR, VO): existing releases
-           that *have* French audio are **flagged** with
-           ``'french_lang_supersede'`` so the dupe checker keeps them as
-           blocking dupes regardless of other exclusion criteria.
-        """
-        upload_audio = await self._build_audio_string(meta)
-
-        # MUET (silent film) — special category, not subject to French lang checks
-        if upload_audio.startswith("MUET"):
-            return dupes
-
-        # Determine the upload's French language level
-        upload_tag, upload_level = self._extract_french_lang_tag(upload_audio)
-        if not upload_tag:
-            # No recognised tag in the audio string — try the raw string
-            # e.g. "MULTI.VFF" → extract "MULTI"
-            for part in upload_audio.split("."):
-                t, lv = self._extract_french_lang_tag(part)
-                if lv > upload_level:
-                    upload_tag, upload_level = t, lv
-
-        # ── Case 1: Upload HAS French audio → drop inferior dupes ──
-        if upload_level >= _FRENCH_AUDIO_THRESHOLD:
-            filtered: list[dict[str, Any]] = []
-            for dupe in dupes:
-                name = dupe.get("name", "") if isinstance(dupe, dict) else str(dupe)
-                _, existing_level = self._extract_french_lang_tag(name)
-                # Keep the dupe only if it also has French audio (or no tag at all,
-                # meaning we can't tell — safer to show it)
-                if existing_level >= _FRENCH_AUDIO_THRESHOLD or existing_level == 0:
-                    filtered.append(dupe)
-                # else: existing is VOSTFR/VO — inferior, silently drop
-            return filtered
-
-        # ── Case 2: Upload LACKS French audio → flag superior dupes ──
-        if upload_audio in ("VOSTFR", "") or upload_level < _FRENCH_AUDIO_THRESHOLD:
-            for dupe in dupes:
-                name = dupe.get("name", "") if isinstance(dupe, dict) else str(dupe)
-                _, existing_level = self._extract_french_lang_tag(name)
-                if existing_level >= _FRENCH_AUDIO_THRESHOLD and isinstance(dupe, dict):
-                    flags: list[str] = dupe.setdefault("flags", [])
-                    if "french_lang_supersede" not in flags:
-                        flags.append("french_lang_supersede")
-
-        return dupes
-
-    async def search_existing(self, meta: Meta, _: Any = None) -> list[dict[str, Any]]:
-        """Wrap the parent's ``search_existing`` with French dupe flagging.
-
-        Trackers that define their *own* ``search_existing`` (C411, V3X)
-        take priority via MRO and call
-        :meth:`_check_french_lang_dupes` explicitly.  This wrapper handles
-        trackers that inherit ``search_existing`` from a parent class
-        (e.g. G3MINI / TOS inheriting from UNIT3D).
-        """
-        dupes = await super().search_existing(meta, _)  # type: ignore[misc]
-        return await self._check_french_lang_dupes(dupes, meta)
-
-    @staticmethod
-    def _detect_truefrench(meta: Meta) -> bool:
-        """Check if the release path/name indicates TRUEFRENCH."""
-        return any("TRUEFRENCH" in str(meta.get(field, "")).upper() for field in ("uuid", "name", "path"))
-
-    @staticmethod
-    def _detect_vfi(meta: Meta) -> bool:
-        """Check if the release path/name indicates VFI."""
-        for field in ("uuid", "name", "path"):
-            val = str(meta.get(field, "")).upper()
-            if re.search(r"[\.\-_]VFI[\.\-_]", val) or val.endswith(".VFI") or val.endswith("-VFI"):
-                return True
-        return False
-
-    @staticmethod
-    def _detect_vfq(meta: Meta) -> bool:
-        """Check if the release path/name indicates VFQ (Québec French)."""
-        for field in ("uuid", "name", "path"):
-            val = str(meta.get(field, "")).upper()
-            if re.search(r"(?:^|[\.\-_\s])VFQ(?:[\.\-_\s]|$)", val):
-                return True
-        return False
-
-    @staticmethod
-    def _detect_vff(meta: Meta) -> bool:
-        """Check if the release path/name indicates VFF (France French)."""
-        for field in ("uuid", "name", "path"):
-            val = str(meta.get(field, "")).upper()
-            if re.search(r"(?:^|[\.\-_\s])VFF(?:[\.\-_\s]|$)", val):
-                return True
-        return False
-
-    @staticmethod
-    def _detect_vf2(meta: Meta) -> bool:
-        """Check if the release path/name indicates VF2 (dual French: VFF + VFQ)."""
-        for field in ("uuid", "name", "path"):
-            val = str(meta.get(field, "")).upper()
-            if re.search(r"(?:^|[\.\-_\s])VF2(?:[\.\-_\s]|$)", val):
-                return True
-        return False
-
-    @staticmethod
-    def _detect_vfb(meta: Meta) -> bool:
-        """Check if the release path/name indicates VFB (Belgian French)."""
-        for field in ("uuid", "name", "path"):
-            val = str(meta.get(field, "")).upper()
-            if re.search(r"(?:^|[\.\-_\s])VFB(?:[\.\-_\s]|$)", val):
-                return True
-        return False
-
-    @staticmethod
-    def _detect_subfrench(meta: Meta) -> bool:
-        """Check if the release path/name indicates SUBFRENCH or VOSTFR.
-
-        Used as a filename-based fallback when MediaInfo does not detect
-        French subtitles (e.g. external .srt files, untagged tracks).
-        """
-        for field in ("uuid", "name", "path"):
-            val = str(meta.get(field, "")).upper()
-            if re.search(r"(?:^|[\.\-_\s])(?:SUBFRENCH|VOSTFR)(?:[\.\-_\s]|$)", val):
-                return True
-        return False
-
-    # ──────────────────────────────────────────────────────────
-    #  Build audio/language string
-    # ──────────────────────────────────────────────────────────
-
-    async def _build_audio_string(self, meta: Meta) -> str:
-        """Build the French language tag from MediaInfo audio tracks.
-
-        Returns one of:
-            Single:  VOF · VFF · VFI · VFQ
-            Multi:   MULTI.VOF · MULTI.VFF · MULTI.VFQ · MULTI.VF2
-            Subs:    VOSTFR
-            Silent:  MUET  (or MUET.VOSTFR)
-            VO:      '' (empty — English or other VO)
-
-        Note: TRUEFRENCH in source filenames is converted to VFF (modern equivalent).
-        """
-        if "mediainfo" not in meta or "media" not in meta.get("mediainfo", {}):
-            return ""
-
-        audio_tracks = self._get_audio_tracks(meta)
-
-        # MUET — MediaInfo present but no audio tracks
-        if not audio_tracks:
-            return "MUET.VOSTFR" if self._has_french_subs(meta) else "MUET"
-
-        ad_audio_tracks = [track for track in audio_tracks if self._is_audio_desc_track(track)]
-        main_audio_tracks = [track for track in audio_tracks if not self._is_audio_desc_track(track)]
-
-        audio_langs = self._extract_audio_languages(main_audio_tracks, meta)
-        if not audio_langs and not ad_audio_tracks:
-            return ""
-
-        has_french_audio = "FRA" in audio_langs
-        has_french_subs = self._has_french_subs(meta)
-        num_audio_tracks = len(main_audio_tracks)
-        fr_suffix = self._get_french_dub_suffix(main_audio_tracks)
-        ad_audio_langs = self._extract_audio_languages(ad_audio_tracks)
-        has_non_french_ad = any(lang != "FRA" for lang in ad_audio_langs)
-        has_audiodesc = bool(meta.get("has_audiodesc") or ad_audio_tracks)
-        is_original_french = str(meta.get("original_language", "")).lower() == "fr"
-        is_truefrench = self._detect_truefrench(meta)
-        is_vfi = self._detect_vfi(meta)
-        is_vfq_filename = self._detect_vfq(meta)
-        is_vfb_filename = self._detect_vfb(meta)
-        is_vff_filename = self._detect_vff(meta)
-        is_vf2_filename = self._detect_vf2(meta)
-
-        def _fr_precision() -> str:
-            """Determine the best French precision tag."""
-            if fr_suffix == "VF2":
-                return "VF2"
-            # VF2 from filename when MediaInfo doesn't have region codes
-            if is_vf2_filename:
-                return "VF2"
-            if is_original_french:
-                # Original French production. Distinguish Québec (fr-CA audio) from
-                # France: a Québécois original is VOQ, a French one VOF. Filename VFQ
-                # only counts as a fallback when MediaInfo carries no explicit region,
-                # so an explicit fr-FR (VFF) is never overridden to VOQ.
-                if fr_suffix == "VFQ" or (not fr_suffix and is_vfq_filename):
-                    return "VOQ"
-                return "VOF"
-            if is_vfi:
-                return "VFI"
-            if fr_suffix == "VFQ":
-                return "VFQ"
-            if fr_suffix == "VFB":
-                return "VFB"
-            if fr_suffix == "VFF":
-                return "VFF"
-            # MediaInfo has generic 'fr' without region — check filename
-            if is_vfq_filename:
-                return "VFQ"
-            if is_vfb_filename:
-                return "VFB"
-            if is_vff_filename or is_truefrench:
-                return "VFF"
-            # Generic 'fr' without region — conservative default
-            return "VFF"
-
-        # ── No French audio ──
-        if not has_french_audio:
-            # MediaInfo subs OR filename hint (SUBFRENCH / VOSTFR)
-            language = "VOSTFR" if has_french_subs or self._detect_subfrench(meta) else ""
-        # ── MULTi — 2+ audio tracks (or non-French track present) ──
-        elif [la for la in audio_langs if la != "FRA"] or num_audio_tracks > 1 or has_non_french_ad:
-            language = f"MULTI.{_fr_precision()}"
-        # ── Single French track ──
-        else:
-            # Includes the original-French case: _fr_precision() returns VOF, or
-            # VOQ when the original audio is Québécois (fr-CA).
-            language = _fr_precision()
-
-        # ── Audio Description prefix ──
-        if language and has_audiodesc and self._should_include_ad_prefix(has_french_audio, ad_audio_langs):
-            language = f"AD.{language}"
-
-        return language
-
-    # ──────────────────────────────────────────────────────────
-    #  French title from TMDB
-    # ──────────────────────────────────────────────────────────
-
-    async def _get_french_title(self, meta: Meta) -> str:
-        """Get French title from TMDB, cached in ``meta['frtitle']``.
-
-        If TMDB returns the original-language title (i.e. no actual French
-        translation exists), falls back to the English title stored in
-        ``meta['title']``.  Exception: if the work is originally French,
-        the original title *is* the French title and is kept.
-        """
-        if meta.get("frtitle"):
-            return meta["frtitle"]
-
-        tmdb_mgr: Any = getattr(self, "tmdb_manager", None)
-        if tmdb_mgr is None:
-            return meta.get("title", "")
-
-        try:
-            fr_data = await tmdb_mgr.get_tmdb_localized_data(meta, data_type="main", language="fr", append_to_response="") or {}
-            fr_title = str(fr_data.get("title", "") or fr_data.get("name", "")).strip()
-            original = str(fr_data.get("original_title", "") or fr_data.get("original_name", "")).strip()
-            orig_lang = str(fr_data.get("original_language", "")).strip().lower()
-            if fr_title and (fr_title != original or orig_lang == "fr"):
-                meta["frtitle"] = fr_title
-                return fr_title
-        except Exception:
-            pass
-
-        return meta.get("title", "")
-
-    # ──────────────────────────────────────────────────────────
-    #  Release naming   (dot-separated, French-tracker convention)
-    #
-    #  Film:  Nom.Année.Edition.Repack.Langue.Résolution.Hybrid.Source.HDR.Audio.Codec-TAG
-    #  TV:    Nom.Année.SXXEXX.Edition.Repack.Langue.Résolution.Hybrid.Source.HDR.Audio.Codec-TAG
-    # ──────────────────────────────────────────────────────────
-
-    async def get_name(self, meta: Meta) -> dict[str, str]:
-        """Build the dot-separated release name (French-tracker conventions)."""
-
-        # When PREFER_ORIGINAL_TITLE is set and the movie is not originally
-        # French, use the original (English) title instead of the TMDB French
-        # translation.  For originally-French works the French title *is* the
-        # original, so we always fetch it.
-        is_original_french = str(meta.get("original_language", "")).lower() == "fr"
-        if self.PREFER_ORIGINAL_TITLE and not is_original_french:
-            title = meta.get("title", "")
-        else:
-            title = await self._get_french_title(meta)
-        language = await self._build_audio_string(meta)
-
-        year = meta.get("year", "")
-        manual_year = meta.get("manual_year")
-        if manual_year is not None and int(manual_year) > 0:
-            year = manual_year
-
-        resolution = meta.get("resolution", "")
-        if resolution == "OTHER":
-            resolution = ""
-        audio = self._get_audio_for_name(meta)
-        service = meta.get("service", "") if self.INCLUDE_SERVICE_IN_NAME else ""
-        season = meta.get("season", "")
-        episode = meta.get("episode", "")
-        part = meta.get("part", "")
-        repack = meta.get("repack", "")
-        three_d = meta.get("3D", "")
-        tag = meta.get("tag", "")
-        source = meta.get("source", "")
-        light = self._light_encode_tag(meta)
-        uhd = meta.get("uhd", "")
-        hdr = meta.get("hdr", "").replace("HDR10+", "HDR10PLUS")
-        hybrid = str(meta.get("webdv", "")) if meta.get("webdv", "") else ""
-        edition = self._format_edition(meta.get("edition", ""))
-        if "hybrid" in edition.upper() or "custom" in edition.upper():
-            edition = re.sub(r"\b(?:Hybrid|CUSTOM|Custom)\b", "", edition, flags=re.IGNORECASE).strip()
-
-        type_val = meta.get("type", "").upper()
-        category = meta.get("category", "MOVIE")
-
-        # Some trackers (e.g. C411) only allow UHD for REMUX/DISC releases
-        if self.UHD_ONLY_FOR_REMUX_DISC and type_val not in ("REMUX", "DISC"):
-            uhd = ""
-
-        video_codec = ""
-        video_encode = ""
-        region = ""
-        dvd_size = ""
-
-        if meta.get("is_disc") == "BDMV":
-            video_codec = meta.get("video_codec", "").replace("H.264", "H264").replace("H.265", "H265").replace("VC-1", "VC1")
-            region = meta.get("region", "") or ""
-        elif meta.get("is_disc") == "DVD":
-            region = meta.get("region", "") or ""
-            dvd_size = meta.get("dvd_size", "")
-        else:
-            video_codec = meta.get("video_codec", "").replace("H.264", "H264").replace("H.265", "H265").replace("VC-1", "VC1")
-            video_encode = meta.get("video_encode", "").replace("H.264", "H264").replace("H.265", "H265").replace("VC-1", "VC1")
-
-        if category == "TV":
-            year = meta["year"] if meta.get("search_year", "") != "" else ""
-            if meta.get("manual_date"):
-                season = ""
-                episode = ""
-        if meta.get("no_season", False) is True:
-            season = ""
-        if meta.get("no_year", False) is True:
-            year = ""
-
-        web_lbl = self.WEB_LABEL  # "WEB" or "WEB-DL" depending on tracker
-
-        # Some trackers want video codec before audio.
-        # Set AUDIO_BEFORE_VIDEO = False on the subclass to flip the order.
-        _audio_first = getattr(self, "AUDIO_BEFORE_VIDEO", True)
-
-        def _av(codec: str, aud: str) -> str:
-            return f"{aud} {codec}" if _audio_first else f"{codec} {aud}"
-
-        name = ""
-
-        # ── MOVIE ──
-        if category == "MOVIE":
-            if type_val == "DISC":
-                disc = meta.get("is_disc", "")
-                if disc == "BDMV":
-                    name = f"{title} {year} {three_d} {edition} {repack} {language} {resolution} {hybrid} {region} {uhd} {source} {hdr} {_av(video_codec, audio)}"
-                elif disc == "DVD":
-                    name = f"{title} {year} {edition} {repack} {language} {region} {source} {dvd_size} {audio}"
-                elif disc == "HDDVD":
-                    name = f"{title} {year} {edition} {repack} {language} {resolution} {source} {_av(video_codec, audio)}"
-            elif type_val == "REMUX" and source in ("BluRay", "HDDVD"):
-                name = f"{title} {year} {three_d} {edition} {repack} {language} {resolution} {hybrid} {uhd} {source} REMUX {hdr} {_av(video_codec, audio)}"
-            elif type_val == "REMUX" and source in ("PAL DVD", "NTSC DVD", "DVD"):
-                name = f"{title} {year} {edition} {repack} {language} {source} REMUX {audio}"
-            elif type_val == "REMUX":
-                name = f"{title} {year} {edition} {repack} {language} {resolution} {hybrid} {uhd} {source} REMUX {hdr} {_av(video_codec, audio)}"
-            elif type_val == "ENCODE":
-                name = f"{title} {year} {edition} {repack} {language} {resolution} {hybrid} {uhd} {source} {light} {hdr} {_av(video_encode, audio)}"
-            elif type_val == "WEBDL":
-                name = f"{title} {year} {edition} {repack} {language} {resolution} {hybrid} {uhd} {service} {web_lbl} {hdr} {_av(video_encode, audio)}"
-            elif type_val == "WEBRIP":
-                name = f"{title} {year} {edition} {repack} {language} {resolution} {hybrid} {uhd} {service} WEBRip {hdr} {_av(video_encode, audio)}"
-            elif type_val == "HDTV":
-                name = f"{title} {year} {edition} {repack} {language} {resolution} {source} {_av(video_encode, audio)}"
-            elif type_val == "DVDRIP":
-                name = f"{title} {year} {repack} {language} {source} DVDRip {_av(video_encode, audio)}"
-
-        # ── TV ──
-        elif category == "TV":
-            se = f"{season}{episode}"
-            if type_val == "DISC":
-                disc = meta.get("is_disc", "")
-                if disc == "BDMV":
-                    name = f"{title} {year} {se} {three_d} {edition} {repack} {language} {resolution} {hybrid} {region} {uhd} {source} {hdr} {_av(video_codec, audio)}"
-                elif disc == "DVD":
-                    name = f"{title} {year} {se} {three_d} {edition} {repack} {language} {region} {source} {dvd_size} {audio}"
-                elif disc == "HDDVD":
-                    name = f"{title} {year} {se} {edition} {repack} {language} {resolution} {source} {_av(video_codec, audio)}"
-            elif type_val == "REMUX" and source in ("BluRay", "HDDVD"):
-                name = f"{title} {year} {se} {part} {three_d} {edition} {repack} {language} {resolution} {hybrid} {uhd} {source} REMUX {hdr} {_av(video_codec, audio)}"
-            elif type_val == "REMUX" and source in ("PAL DVD", "NTSC DVD", "DVD"):
-                name = f"{title} {year} {se} {part} {edition} {repack} {language} {source} REMUX {audio}"
-            elif type_val == "REMUX":
-                name = f"{title} {year} {se} {part} {edition} {repack} {language} {resolution} {hybrid} {uhd} {source} REMUX {hdr} {_av(video_codec, audio)}"
-            elif type_val == "ENCODE":
-                name = f"{title} {year} {se} {part} {edition} {repack} {language} {resolution} {hybrid} {uhd} {source} {light} {hdr} {_av(video_encode, audio)}"
-            elif type_val == "WEBDL":
-                name = f"{title} {year} {se} {part} {edition} {repack} {language} {resolution} {hybrid} {uhd} {service} {web_lbl} {hdr} {_av(video_encode, audio)}"
-            elif type_val == "WEBRIP":
-                name = f"{title} {year} {se} {part} {edition} {repack} {language} {resolution} {hybrid} {uhd} {service} WEBRip {hdr} {_av(video_encode, audio)}"
-            elif type_val == "HDTV":
-                name = f"{title} {year} {se} {part} {edition} {repack} {language} {resolution} {source} {_av(video_encode, audio)}"
-            elif type_val == "DVDRIP":
-                name = f"{title} {year} {se} {repack} {language} {source} DVDRip {_av(video_encode, audio)}"
-
-        if not name:
-            name = f"{title} {year} {language} {resolution} {type_val} {_av(video_encode, audio)}"
-
-        # ── Post-processing ──
-        name = " ".join(name.split())  # collapse whitespace
-
-        # Handle notag: if tag is empty/invalid and tracker accepts notag, use the label
-        tag_group = tag.strip("-").strip().lower() if tag else ""
-        invalid_tags = ["nogrp", "nogroup", "unknown", "unk"]
-        if not tag_group or any(inv == tag_group for inv in invalid_tags):
-            label = getattr(self, "notag_label", "")
-            if label:
-                tag = f"-{label}"
-
-        name = name + tag  # tag starts with '-', no space needed
-
-        return self._format_name(name)
-
-    @staticmethod
-    def _normalize_audio_name_tokens(dot_name: str) -> str:
-        """Normalize audio codec tokens in a dotted release name to the
-        French-scene convention: DD→AC3, TRUEHD, DTS.HD.MA, DTS.X, and
-        ATMOS placed between the codec and the channel count.
-        """
-        # DD → AC3 (but not DDP which stays as-is)
-        dot_name = re.sub(r"\.DD\.", ".AC3.", dot_name)
-        # TrueHD → TRUEHD (case normalization)
-        dot_name = re.sub(r"\.TrueHD\.", ".TRUEHD.", dot_name, flags=re.IGNORECASE)
-        dot_name = re.sub(r"\.TrueHD$", ".TRUEHD", dot_name, flags=re.IGNORECASE)
-        # DTS-HD.MA → DTS.HD.MA (dash to dot)
-        dot_name = dot_name.replace(".DTS-HD.MA.", ".DTS.HD.MA.")
-        dot_name = dot_name.replace(".DTS-HD.HRA.", ".DTS.HD.HRA.")
-        # DTS:X → DTS.X (colon to dot)
-        dot_name = dot_name.replace(".DTS:X.", ".DTS.X.")
-        dot_name = dot_name.replace(".DTSX.", ".DTS.X.")
-        # Atmos capitalization
-        dot_name = re.sub(r"\.Atmos\.", ".ATMOS.", dot_name, flags=re.IGNORECASE)
-        dot_name = re.sub(r"\.Atmos$", ".ATMOS", dot_name, flags=re.IGNORECASE)
-        # ATMOS must appear AFTER the audio codec and BEFORE audio channels : DDP.5.1.ATMOS → DDP.ATMOS.5.1
-        # Pattern 1: codec.channels.ATMOS → codec.ATMOS.channels
-        dot_name = re.sub(r"\.(DDP|AC3|EAC3|DTS|TRUEHD|FLAC|AAC|LPCM|DTS\.HD\.MA|DTS\.HD\.HRA|DTS\.X)\.(\d\.\d)\.ATMOS([.-])", r".\1.ATMOS.\2\3", dot_name, flags=re.IGNORECASE)
-        # Pattern 2: ATMOS.codec.channels → codec.ATMOS.channels
-        dot_name = re.sub(r"\.ATMOS\.(DDP|AC3|EAC3|DTS|TRUEHD|FLAC|AAC|LPCM|DTS\.HD\.MA|DTS\.HD\.HRA|DTS\.X)\.(\d\.\d)([.-])", r".\1.ATMOS.\2\3", dot_name, flags=re.IGNORECASE)
-        return dot_name
-
-    @staticmethod
-    def _enforce_web_codec_convention(meta: Meta, name: str) -> str:
-        """WEB codec token per the actual MediaInfo, not the type label:
-        untouched WEB streams (no Encoded_Library_Settings) use H264/H265,
-        re-encodes use x264/x265.
-        """
-        if str(meta.get("type", "")).upper() in ("WEBDL", "WEBRIP"):
-            if meta.get("has_encode_settings", False):
-                name = re.sub(r"\.H264\b", ".x264", name, flags=re.IGNORECASE)
-                name = re.sub(r"\.H265\b", ".x265", name, flags=re.IGNORECASE)
-            else:
-                name = re.sub(r"\.x264\b", ".H264", name, flags=re.IGNORECASE)
-                name = re.sub(r"\.x265\b", ".H265", name, flags=re.IGNORECASE)
-        return name
-
-    def _format_name(self, raw_name: str) -> dict[str, str]:
-        """Clean and format the release name (dot-separated by default).
-
-        Subclasses may override this to change the separator (e.g. spaces).
-        """
-        clean = self._fr_clean(raw_name)
-        dot_name = clean.replace(" ", ".")
-
-        # Keep only the LAST hyphen (group-tag separator)
-        idx = dot_name.rfind("-")
-        if idx > 0:
-            dot_name = dot_name[:idx].replace("-", ".") + dot_name[idx:]
-
-        # Remove isolated hyphens between dots
-        dot_name = re.sub(r"\.(-\.)+", ".", dot_name)
-        # Collapse consecutive dots, strip boundary dots
-        dot_name = re.sub(r"\.{2,}", ".", dot_name).strip(".")
-
-        return {"name": dot_name}
-
-    # Map special Unicode chars to their ASCII equivalents *before*
-    # unidecode (which would map · → * and lose the separator).
-    _TITLE_CHAR_MAP: dict[str, str] = {
-        "\u00b7": " ",  # middle dot   (WALL·E → WALL E → WALL.E / Wall E)
-        "\u2022": " ",  # bullet       (same rationale)
-        "\u2010": "-",  # hyphen
-        "\u2011": "-",  # non-breaking hyphen
-        "\u2012": "-",  # figure dash
-        "\u2013": "-",  # en dash
-        "\u2014": "-",  # em dash
-        "\u2015": "-",  # horizontal bar
-        "\u2212": "-",  # minus sign
-    }
-
-    @staticmethod
-    def _fr_clean(text: str) -> str:
-        """Strip accents and non-filename characters.
-
-        French elided articles (l', d', qu', etc.) are expanded so that
-        the apostrophe becomes a space while preserving the original case:
-        ``l'autre`` → ``l autre``, ``L'Ordre`` → ``L Ordre``.
-        """
-        import unicodedata
-
-        # Normalise to NFC first so that combining-character sequences
-        # (e.g. U+0065 U+0301 for "é") are collapsed into their
-        # precomposed form before unidecode sees them.  Without this,
-        # unidecode turns the bare combining accent (U+0301) into an
-        # empty string, silently dropping the following vowel.
-        text = unicodedata.normalize("NFC", text)
-        for char, repl in FrenchTrackerMixin._TITLE_CHAR_MAP.items():
-            text = text.replace(char, repl)
-        text = unidecode(text)
-        # Replace apostrophes / RIGHT SINGLE QUOTATION MARK / backticks
-        # that follow a French elided article with a space, preserving
-        # the original case:  l'autre → l autre,  L'Ordre → L Ordre
-        text = re.sub(
-            r"\b([lLdDnNsScCjJmM]|[Qq]u|[Jj]usqu|[Ll]orsqu|[Pp]uisqu)['\u2019`]",
-            lambda m: m.group(1) + " ",
-            text,
-        )
-        return re.sub(r"[^a-zA-Z0-9 .+\-]", "", text)
-
-    # ──────────────────────────────────────────────────────────
-    #  MediaInfo parsing helpers (shared by description builders)
-    # ──────────────────────────────────────────────────────────
-
-    @staticmethod
-    def _lang_to_flag(lang: str) -> str:
-        """Map a language name (from MediaInfo) to its flag emoji."""
-        key = lang.lower().split("(")[0].strip()
-        return LANG_FLAGS.get(key, "\U0001f3f3\ufe0f")
-
-    @staticmethod
-    def _lang_to_french_name(lang: str) -> str:
-        """Map a language name (from MediaInfo) to its French display name."""
-        key = lang.lower().split("(")[0].strip()
-        return LANG_NAMES_FR.get(key, lang)
-
-    @staticmethod
-    def _channels_to_layout(channels: str) -> str:
-        """Convert MI channel count to layout notation.
-
-        '6 channels' → '5.1', '8 channels' → '7.1', '2 channels' → '2.0', etc.
-        """
-        m = re.search(r"(\d+)", channels)
-        if not m:
-            return channels
-        n = int(m.group(1))
-        mapping = {1: "1.0", 2: "2.0", 3: "2.1", 6: "5.1", 8: "7.1"}
-        return mapping.get(n, str(n))
-
-    @staticmethod
-    def _parse_mi_audio_tracks(mi_text: str) -> list[dict[str, str]]:
-        """Parse audio tracks from MediaInfo text into structured dicts.
-
-        Each dict may contain: language, format, commercial_name, bitrate,
-        channels, channel_layout, title.
-        """
-        tracks: list[dict[str, str]] = []
-        if not mi_text:
-            return tracks
-        current: Optional[dict[str, str]] = None
-
-        for line in mi_text.split("\n"):
-            stripped = line.strip()
-            if stripped == "Audio" or stripped.startswith("Audio #"):
-                if current:
-                    tracks.append(current)
-                current = {}
-                continue
-            if current is not None and (
-                stripped.startswith("Text") or stripped.startswith("Menu") or stripped == "Video" or stripped.startswith("Video #") or stripped == "General"
-            ):
-                tracks.append(current)
-                current = None
-            if current is not None and ":" in stripped:
-                key, _, val = stripped.partition(":")
-                key = key.strip()
-                val = val.strip()
-                if key == "Language":
-                    current["language"] = val
-                elif key == "Format":
-                    current["format"] = val
-                elif key == "Commercial name":
-                    current["commercial_name"] = val
-                elif key == "Bit rate":
-                    current["bitrate"] = val
-                elif key == "Channel(s)":
-                    current["channels"] = val
-                elif key == "Channel layout":
-                    current["channel_layout"] = val
-                elif key == "Title":
-                    current["title"] = val
-                elif key == "Default":
-                    current["default"] = val
-
-        if current:
-            tracks.append(current)
-        return tracks
-
-    @staticmethod
-    def _parse_mi_subtitle_tracks(mi_text: str) -> list[dict[str, str]]:
-        """Parse subtitle tracks from MediaInfo text into structured dicts.
-
-        Each dict may contain: language, format, title, forced, default.
-        """
-        tracks: list[dict[str, str]] = []
-        if not mi_text:
-            return tracks
-        current: Optional[dict[str, str]] = None
-
-        for line in mi_text.split("\n"):
-            stripped = line.strip()
-            if stripped == "Text" or stripped.startswith("Text #"):
-                if current:
-                    tracks.append(current)
-                current = {}
-                continue
-            if current is not None and (stripped.startswith("Menu") or stripped.startswith("Audio") or stripped == "Video" or stripped == "General"):
-                tracks.append(current)
-                current = None
-            if current is not None and ":" in stripped:
-                key, _, val = stripped.partition(":")
-                key = key.strip()
-                val = val.strip()
-                if key == "Language":
-                    current["language"] = val
-                elif key == "Format":
-                    current["format"] = val
-                elif key == "Title":
-                    current["title"] = val
-                elif key == "Forced":
-                    current["forced"] = val
-                elif key == "Default":
-                    current["default"] = val
-                elif key == "Count of elements":
-                    current["element_count"] = val
-
-        if current:
-            tracks.append(current)
-        return tracks
-
-    @staticmethod
-    def _sub_format_short(fmt: str) -> str:
-        """Return a short label for a subtitle format string."""
-        up = fmt.upper()
-        if "PGS" in up:
-            return "PGS"
-        if "SRT" in up or "UTF-8" in up:
-            return "SRT"
-        if "ASS" in up or "SSA" in up:
-            return "ASS"
-        if "VOBSUB" in up:
-            return "VobSub"
-        return fmt
-
-    # ── Release type labels ───────────────────────────────────────────
-    TYPE_LABELS: dict[str, str] = {
-        "DISC": "Disc",
-        "REMUX": "Remux",
-        "ENCODE": "Encode",
-        "WEBDL": "WEB-DL",
-        "WEBRIP": "WEBRip",
-        "HDTV": "HDTV",
-        "DVDRIP": "DVDRip",
-    }
-
-    @staticmethod
-    def _get_type_label(meta: dict) -> str:
-        """Return a human-readable release type label."""
-        raw = (meta.get("type") or "").upper()
-        return FrenchTrackerMixin.TYPE_LABELS.get(raw, raw)
-
-    @staticmethod
-    def _light_encode_tag(meta: dict) -> str:
-        """French-scene light re-encode label, or '' when not one.
-
-        4KLight (2160p) and HDLight (1080p) are BluRay re-encodes; the source
-        release name declares which, but it isn't a meta field, so we read it
-        from ``uuid`` exactly like the C411 slot/quality detection does. Surfaced
-        in the generated name (after the source) so the tag isn't dropped.
-        """
-        uuid = str(meta.get("uuid", "")).lower()
-        if "4klight" in uuid:
-            return "4KLight"
-        if "hdlight" in uuid:
-            return "HDLight"
-        return ""
-
-    # Container name → common file extension
-    CONTAINER_EXT: dict[str, str] = {
-        "MATROSKA": "MKV",
-        "AVI": "AVI",
-        "MPEG-4": "MP4",
-        "MPEG-TS": "TS",
-        "BDAV": "M2TS",
-        "WEBM": "WEBM",
-        "OGG": "OGG",
-        "FLASH VIDEO": "FLV",
-        "WINDOWS MEDIA": "WMV",
-    }
-
-    @staticmethod
-    def _parse_mi_container(mi_text: str) -> str:
-        """Extract container format from the MI General section."""
-        if not mi_text:
-            return ""
-        for line in mi_text.split("\n"):
-            stripped = line.strip()
-            if stripped.startswith("Format") and ":" in stripped and "profile" not in stripped.lower():
-                match = re.search(r":\s*(.+)", stripped)
-                if match:
-                    return match.group(1).strip()
-            # Stop after General section
-            if stripped in ("Video", "Audio", "Text", "Menu") or stripped.startswith("Video"):
-                break
-        return ""
-
-    @classmethod
-    def _format_container(cls, mi_text: str) -> str:
-        """Return container with file extension, e.g. 'MATROSKA (MKV)'."""
-        raw = cls._parse_mi_container(mi_text)
-        if not raw:
-            return ""
-        upper = raw.upper()
-        ext = cls.CONTAINER_EXT.get(upper, "")
-        return f"{upper} ({ext})" if ext else upper
-
-    @staticmethod
-    def _get_release_group(meta: dict) -> str:
-        """Extract release group name from meta['tag'] (strip leading hyphen)."""
-        tag = (meta.get("tag") or "").strip()
-        return tag.lstrip("-") if tag else ""
-
-    # ── Total size / file count (season packs vs single files) ──────────
-
-    @staticmethod
-    def _get_total_size(meta: dict, mi_text: str) -> str:
-        """Return human-readable total size for the release.
-
-        For a single file, use the MediaInfo 'File size' line.
-        For a directory (season pack), sum every file on disk.
-        """
-        path = meta.get("path", "")
-        if path and os.path.isdir(path):
-            total = sum(os.path.getsize(os.path.join(root, f)) for root, _dirs, files in os.walk(path) for f in files)
-            if total <= 0:
-                return ""
-            # Format to GiB / MiB like MediaInfo does
-            if total >= 1 << 30:  # >= 1 GiB
-                return f"{total / (1 << 30):.2f} GiB"
-            if total >= 1 << 20:  # >= 1 MiB
-                return f"{total / (1 << 20):.2f} MiB"
-            return f"{total / (1 << 10):.2f} KiB"
-        # Single file: use MediaInfo
-        if mi_text:
-            size_match = re.search(r"File size\s*:\s*(.+?)\s*(?:\n|$)", mi_text)
-            if size_match:
-                return size_match.group(1).strip()
-        return ""
-
-    # Extensions included in the torrent (video files).
-    _TORRENT_EXTENSIONS: frozenset[str] = frozenset((".mkv", ".mp4", ".ts", ".m2ts", ".vob", ".avi"))
-
-    def _count_files(self, meta: dict) -> str:
-        """Count files actually included in the torrent.
-
-        Only video extensions are counted (matching torrent creation logic
-        which excludes .nfo, .jpg, .srt, etc.).
-        """
-        path = meta.get("path", "")
-        if not path or not os.path.exists(path):
-            return ""
-        if os.path.isfile(path):
-            return "1"
-        exts = self._TORRENT_EXTENSIONS
-        count = sum(1 for _, _, files in os.walk(path) for f in files if os.path.splitext(f)[1].lower() in exts)
-        return str(count) if count else ""
-
-    # ── HDR / Dolby Vision display (plain text labels) ──────────────────
-    HDR_LABELS: dict[str, str] = {
-        "DV": "Dolby Vision",
-        "HDR10+": "HDR10+",
-        "HDR": "HDR10",
-        "HLG": "HLG",
-        "PQ10": "PQ10",
-        "WCG": "WCG",
-    }
-
-    def _format_hdr_dv_bbcode(self, meta: dict) -> Optional[str]:
-        """Return a plain-text string listing HDR formats.
-
-        When Dolby Vision is detected, the DV profile (e.g. "Profile 8.1")
-        is appended if available in the MediaInfo JSON data.
-
-        Returns *None* when there is nothing to display (SDR content).
-        """
-        hdr_raw: str = (meta.get("hdr") or "").strip()
-        if not hdr_raw:
-            return None
-
-        # Match longest tokens first so "HDR10+" is not consumed by "HDR".
-        ordered_keys = ["HDR10+", "DV", "HDR", "HLG", "PQ10", "WCG"]
-        remaining = hdr_raw
-        labels: list[str] = []
-        for key in ordered_keys:
-            if key in remaining:
-                label = self.HDR_LABELS[key]
-                # Enrich "Dolby Vision" with the DV profile from MediaInfo JSON
-                if key == "DV":
-                    dv_profile = self._get_dv_profile(meta)
-                    if dv_profile:
-                        label = f"{label} ({dv_profile})"
-                labels.append(label)
-                remaining = remaining.replace(key, "", 1).strip()
-
-        return " + ".join(labels) if labels else None
-
-    @staticmethod
-    def _get_dv_profile(meta: dict) -> str:
-        """Extract a human-readable Dolby Vision profile from MediaInfo JSON.
-
-        ``HDR_Format_Profile`` typically looks like ``dvhe.08.06`` (Profile 8,
-        Level 6) or ``dvhe.05.06``.  We parse it into ``Profile 8.6`` etc.
-        Returns an empty string when unavailable.
-        """
-        tracks = meta.get("mediainfo", {}).get("media", {}).get("track", [])
-        for track in tracks:
-            if track.get("@type") != "Video":
-                continue
-            raw = track.get("HDR_Format_Profile", "")
-            if not raw or isinstance(raw, dict):
-                # Also try HDR_Format_String which may contain "Profile X.Y"
-                hdr_str = track.get("HDR_Format_String", "")
-                if isinstance(hdr_str, str) and "Profile" in hdr_str:
-                    import re as _re
-
-                    m = _re.search(r"Profile\s+(\d+(?:\.\d+)?)", hdr_str)
-                    if m:
-                        return f"Profile {m.group(1)}"
-                return ""
-            # Parse "dvhe.08.06" → Profile 8.6
-            # Format: dvhe.PP.LL or dvav.PP.LL (PP=profile, LL=level)
-            if isinstance(raw, str):
-                import re as _re
-
-                m = _re.search(r"(?:dvhe|dvav)\.(\d+)\.(\d+)", raw)
-                if m:
-                    profile = int(m.group(1))
-                    level = int(m.group(2))
-                    return f"Profile {profile}.{level}"
-                # Fallback: sometimes it's just "dvhe.08"
-                m = _re.search(r"(?:dvhe|dvav)\.(\d+)", raw)
-                if m:
-                    return f"Profile {int(m.group(1))}"
-            break
-        return ""
-
-    def _format_audio_bbcode(self, mi_text: str, meta: Optional[Meta] = None) -> list[str]:
-        """Build pretty BBCode lines for audio tracks.
-
-        When *meta* is provided, cross-references the JSON MediaInfo data
-        (which contains raw BCP-47 language codes like ``fr-CA``) with the
-        text-parsed tracks for accurate VFF/VFQ/VFB detection.
-
-        Detection priority:
-          1. JSON MediaInfo language code (``fr-FR`` → VFF, ``fr-CA`` → VFQ, ``fr-BE`` → VFB)
-          2. Explicit label in the track Title field (VFF, VFQ, VFB, VF2, VOF, VFI)
-          3. No variant suffix — just "Français"
-
-        Returns a list like::
-
-            ['🇫🇷 Français VFF [5.1] : DTS-HD @ 2 046 kb/s',
-             '🇨🇦 Français VFQ [5.1] : Dolby Digital Plus @ 1 024 kb/s',
-             '🇧🇪 Français VFB [5.1] : AC3 @ 448 kb/s',
-             '🇺🇸 Anglais [5.1] : AC3 @ 384 kb/s']
-        """
-        tracks = self._parse_mi_audio_tracks(mi_text)
-        lines: list[str] = []
-
-        # ── Build a list of raw language codes from JSON MediaInfo ──
-        # This lets us detect fr-CA (VFQ) vs fr-FR (VFF) vs fr-BE (VFB)
-        # reliably, because MediaInfo text output only shows "French" for all.
-        json_audio_langs: list[str] = []
-        if meta:
-            try:
-                json_tracks = meta.get("mediainfo", {}).get("media", {}).get("track", [])
-                json_audio_langs = [str(t.get("Language", "")).lower().strip() for t in json_tracks if t.get("@type") == "Audio"]
-            except (AttributeError, TypeError):
-                pass
-
-        default_found = False
-        for i, at in enumerate(tracks):
-            lang = at.get("language", "Unknown")
-            flag = self._lang_to_flag(lang)
-            name = self._lang_to_french_name(lang)
-            channels = at.get("channels", "")
-            layout = self._channels_to_layout(channels) if channels else ""
-            commercial = at.get("commercial_name", "")
-            fmt = at.get("format", "")
-            bitrate = at.get("bitrate", "")
-            title = at.get("title", "").upper()
-
-            # Normalise language: "French (CA)" → base="french", region="ca"
-            lang_lower = lang.lower().strip()
-            region_match = re.search(r"\((\w+)\)", lang_lower)
-            lang_region = region_match.group(1) if region_match else ""
-            lang_base = lang_lower.split("(")[0].strip()
-
-            # For French tracks: detect VFQ/VFF/VFB variant
-            if lang_base in ("french", "fre", "fra", "français", "francais"):
-                variant_detected = False
-
-                # Priority 1: raw BCP-47 language code from JSON MediaInfo
-                if i < len(json_audio_langs):
-                    raw_code = json_audio_langs[i]
-                    if raw_code in ("fr-ca", "fr-qc"):
-                        flag = "🇨🇦"
-                        name = "Français VFQ"
-                        variant_detected = True
-                    elif raw_code == "fr-be":
-                        flag = "🇧🇪"
-                        name = "Français VFB"
-                        variant_detected = True
-                    elif raw_code in ("fr-fr", "fr-ch"):
-                        # VFF / standard France French — "Français" suffices
-                        variant_detected = True
-
-                # Priority 2: region from MI text, e.g. "French (CA)" → "ca"
-                if not variant_detected and lang_region:
-                    if lang_region == "ca":
-                        flag = "🇨🇦"
-                        name = "Français VFQ"
-                        variant_detected = True
-                    elif lang_region == "be":
-                        flag = "🇧🇪"
-                        name = "Français VFB"
-                        variant_detected = True
-                    elif lang_region in ("fr", "ch"):
-                        variant_detected = True
-
-                # Priority 3: explicit label in the track Title field
-                if not variant_detected:
-                    if "VFQ" in title or "QUÉB" in title or "QUEB" in title:
-                        flag = "🇨🇦"
-                        name = "Français VFQ"
-                    elif "VFB" in title or "BELG" in title:
-                        flag = "🇧🇪"
-                        name = "Français VFB"
-                    elif "VFI" in title:
-                        name = "Français VFI"
-                    # VFF, TRUEFRENCH, VOF → just "Français" (default)
-
-            # ── Spanish region detection ──
-            elif lang_base in ("spanish", "spa", "español", "espanol"):
-                variant_detected = False
-
-                if i < len(json_audio_langs):
-                    raw_code = json_audio_langs[i]
-                    if raw_code == "es-es":
-                        flag = "🇪🇸"
-                        variant_detected = True
-                    elif raw_code.startswith("es-") and raw_code != "es-es":
-                        flag = "🇲🇽"
-                        variant_detected = True
-
-                if not variant_detected and lang_region:
-                    if lang_region == "es":
-                        flag = "🇪🇸"
-                    elif lang_region in ("419", "mx", "ar", "co", "cl", "pe", "ve") or "latin" in lang_lower:
-                        flag = "🇲🇽"
-
-                if not variant_detected and not lang_region and title:
-                    if "LATIN" in title or "LATINO" in title:
-                        flag = "🇲🇽"
-                    elif "SPAIN" in title or "ESPAÑA" in title or "CASTILL" in title:
-                        flag = "🇪🇸"
-
-            # ── Portuguese region detection ──
-            elif lang_base in ("portuguese", "por", "português", "portugues"):
-                variant_detected = False
-
-                if i < len(json_audio_langs):
-                    raw_code = json_audio_langs[i]
-                    if raw_code in ("pt-br",):
-                        flag = "🇧🇷"
-                        variant_detected = True
-                    elif raw_code in ("pt-pt", "pt"):
-                        flag = "🇵🇹"
-                        variant_detected = True
-
-                if not variant_detected and lang_region:
-                    if lang_region == "br":
-                        flag = "🇧🇷"
-                    elif lang_region in ("pt",):
-                        flag = "🇵🇹"
-
-                if not variant_detected and not lang_region and title and ("BRAZIL" in title or "BRASIL" in title):
-                    flag = "🇧🇷"
-
-            # ── Mandarin script variant detection ──
-            elif lang_base in ("mandarin",):
-                flag = "🇨🇳"
-                if lang_region == "hant":
-                    name = "Mandarin (traditionnel)"
-                elif lang_region == "hans":
-                    name = "Mandarin (simplifié)"
-
-            # ── Cantonese script variant detection ──
-            elif lang_base in ("cantonese",):
-                flag = "🇨🇳"
-                if lang_region == "hant":
-                    name = "Cantonais (traditionnel)"
-                elif lang_region == "hans":
-                    name = "Cantonais (simplifié)"
-
-            # ── Default track detection ──
-            is_default = at.get("default", "").lower() == "yes" and not default_found
-            if is_default:
-                default_found = True
-
-            # ── Audio Description detection ──
-            is_audio_desc = self._is_audio_desc_track(at)
-
-            # ── Commentary detection ──
-            commentary_tag = ""
-            title_original = at.get("title", "")
-            if title and "COMMENTARY" in title:
-                # Extract short descriptor from title patterns:
-                #   "English [Philosopher Commentary]" → "Philosopher"
-                #   "Cast and Crew Commentary" → "Cast and Crew"
-                #   "Composer Commentary/Music-Only Track" → "Composer"
-                #   "Commentary by Director ..." → too long, just [Commentaire]
-                label = ""
-                # Pattern: "Language [Descriptor Commentary...]"
-                bracket_match = re.search(r"\[([^\]]*commentary[^\]]*)\]", title_original, re.IGNORECASE)
-                if bracket_match:
-                    inner = bracket_match.group(1).strip()
-                    # Remove "Commentary" and anything after "/" from inner text
-                    inner = re.sub(r"\s*Commentary.*", "", inner, flags=re.IGNORECASE).strip()
-                    if inner and inner.lower() != lang_base:
-                        label = inner
-                else:
-                    # Pattern: "Descriptor Commentary" (no brackets)
-                    comm_match = re.match(r"^(.+?)\s+Commentary", title_original, re.IGNORECASE)
-                    if comm_match:
-                        label = comm_match.group(1).strip()
-
-                commentary_tag = f"Commentaire : {label}" if label and len(label) <= 40 else "Commentaire"
-
-            # Build: flag Name [layout] : Codec @ Bitrate
-            parts: list[str] = [f"{flag} {name}"]
-            if is_audio_desc:
-                parts.append(" [AD]")
-            if commentary_tag:
-                parts.append(f" [{commentary_tag}]")
-            if is_default:
-                parts.append(" (piste par défaut)")
-            if layout:
-                parts.append(f" [{layout}]")
-            codec = commercial or fmt
-            if codec:
-                parts.append(f" : {codec}")
-            if bitrate:
-                parts.append(f" @ {bitrate}")
-            lines.append("".join(parts))
-        return lines
-
-    def _format_subtitle_bbcode(self, mi_text: str, meta: Optional[Meta] = None) -> list[str]:
-        """Build pretty BBCode lines for subtitle tracks.
-
-        When *meta* is provided, cross-references the JSON MediaInfo data
-        (which contains raw BCP-47 language codes like ``fr-CA``) with the
-        text-parsed tracks for accurate region flag detection.
-
-        Detection priority (same as audio):
-          1. JSON MediaInfo language code (``fr-FR`` → 🇫🇷, ``fr-CA`` → 🇨🇦, ``es-419`` → 🌎)
-          2. Region from MI text, e.g. ``French (CA)``
-          3. Explicit label in the track Title field
-
-        Returns a list like:
-          ['🇫🇷 Français : PGS (complets)',
-           '🇨🇦 Français : PGS (forcés)',
-           '🇺🇸 Anglais : PGS (SDH)']
-        """
-        tracks = self._parse_mi_subtitle_tracks(mi_text)
-        lines: list[str] = []
-
-        # ── Build a list of raw language codes from JSON MediaInfo ──
-        json_text_langs: list[str] = []
-        if meta:
-            try:
-                json_tracks = meta.get("mediainfo", {}).get("media", {}).get("track", [])
-                json_text_langs = [str(t.get("Language", "")).lower().strip() for t in json_tracks if t.get("@type") == "Text"]
-            except (AttributeError, TypeError):
-                pass
-
-        default_found = False
-        for i, st in enumerate(tracks):
-            lang = st.get("language", "") or "Unknown"
-            flag = self._lang_to_flag(lang)
-            name = self._lang_to_french_name(lang)
-            fmt = st.get("format", "")
-            fmt_short = self._sub_format_short(fmt) if fmt else ""
-            forced = st.get("forced", "").lower() == "yes"
-            is_default = st.get("default", "").lower() == "yes" and not default_found
-            if is_default:
-                default_found = True
-            element_count = st.get("element_count", "")
-            title = st.get("title", "")
-
-            # Detect forced from title field too
-            if not forced and title and "forced" in title.lower():
-                forced = True
-
-            # Detect SDH from title
-            sdh = bool(title and ("sdh" in title.lower() or "hearing" in title.lower()))
-
-            # Normalise language: "French (CA)" → base="french", region="ca"
-            lang_lower = lang.lower().strip()
-            region_match = re.search(r"\((\w+)\)", lang_lower)
-            lang_region = region_match.group(1) if region_match else ""
-            lang_base = lang_lower.split("(")[0].strip()
-
-            # ── French region detection (VFQ / VFB) ──
-            if lang_base in ("french", "fre", "fra", "français", "francais"):
-                variant_detected = False
-
-                # Priority 1: raw BCP-47 language code from JSON MediaInfo
-                if i < len(json_text_langs):
-                    raw_code = json_text_langs[i]
-                    if raw_code in ("fr-ca", "fr-qc"):
-                        flag = "🇨🇦"
-                        variant_detected = True
-                    elif raw_code == "fr-be":
-                        flag = "🇧🇪"
-                        variant_detected = True
-                    elif raw_code in ("fr-fr", "fr-ch"):
-                        variant_detected = True
-
-                # Priority 2: region from MI text, e.g. "French (CA)" → "ca"
-                if not variant_detected and lang_region:
-                    if lang_region == "ca":
-                        flag = "🇨🇦"
-                        variant_detected = True
-                    elif lang_region == "be":
-                        flag = "🇧🇪"
-                        variant_detected = True
-                    elif lang_region in ("fr", "ch"):
-                        variant_detected = True
-
-                # Priority 3: explicit label in the track Title field
-                if not variant_detected and title:
-                    title_upper = title.upper()
-                    if "CANADA" in title_upper or "VFQ" in title_upper or "QUÉB" in title_upper or "QUEB" in title_upper:
-                        flag = "🇨🇦"
-                    elif "BELG" in title_upper or "VFB" in title_upper:
-                        flag = "🇧🇪"
-
-            # ── Spanish region detection ──
-            elif lang_base in ("spanish", "spa", "español", "espanol"):
-                variant_detected = False
-
-                if i < len(json_text_langs):
-                    raw_code = json_text_langs[i]
-                    if raw_code == "es-es":
-                        flag = "🇪🇸"
-                        variant_detected = True
-                    elif raw_code.startswith("es-") and raw_code != "es-es":
-                        # Latin American variant (es-419, es-MX, etc.)
-                        flag = "🇲🇽"
-                        variant_detected = True
-
-                if not variant_detected and lang_region:
-                    if lang_region == "es":
-                        flag = "🇪🇸"
-                    elif lang_region in ("419", "mx", "ar", "co", "cl", "pe", "ve") or "latin" in lang_lower:
-                        flag = "🇲🇽"
-
-                if not variant_detected and not lang_region and title:
-                    title_lower = title.lower()
-                    if "latin" in title_lower or "latino" in title_lower:
-                        flag = "🇲🇽"
-                    elif "spain" in title_lower or "españa" in title_lower or "castill" in title_lower:
-                        flag = "🇪🇸"
-
-            # ── Portuguese region detection ──
-            elif lang_base in ("portuguese", "por", "português", "portugues"):
-                if i < len(json_text_langs):
-                    raw_code = json_text_langs[i]
-                    if raw_code in ("pt-br",):
-                        flag = "🇧🇷"
-                    elif raw_code in ("pt-pt", "pt"):
-                        flag = "🇵🇹"
-                elif lang_region:
-                    if lang_region == "br":
-                        flag = "🇧🇷"
-                elif title:
-                    title_lower = title.lower()
-                    if "brazil" in title_lower or "brasil" in title_lower:
-                        flag = "🇧🇷"
-
-            # ── Mandarin script variant detection ──
-            elif lang_base in ("mandarin",):
-                flag = "🇨🇳"
-                if lang_region == "hant":
-                    name = "Mandarin (traditionnel)"
-                elif lang_region == "hans":
-                    name = "Mandarin (simplifié)"
-
-            # ── Cantonese script variant detection ──
-            elif lang_base in ("cantonese",):
-                flag = "🇨🇳"
-                if lang_region == "hant":
-                    name = "Cantonais (traditionnel)"
-                elif lang_region == "hans":
-                    name = "Cantonais (simplifié)"
-
-            # ── Commentary detection ──
-            is_commentary = bool(title and "commentary" in title.lower())
-
-            # Build qualifier
-            if forced:
-                qualifier = "forcés"
-            elif sdh:
-                qualifier = "SDH"
-            else:
-                qualifier = "complets"
-
-            if is_commentary:
-                qualifier += ", commentaire"
-
-            # Default / forced status indicator
-            if is_default and forced:
-                status = " (piste par défaut et forcée)"
-            elif is_default:
-                status = " (piste par défaut)"
-            elif forced:
-                status = " (piste forcée)"
-            else:
-                status = ""
-
-            # Element count display
-            count_part = f", {element_count} éléments" if element_count else ""
-
-            # Parenthesized info after format: (qualifier, N éléments)
-            paren_inner = f"{qualifier}{count_part}"
-
-            parts: list[str] = [f"{flag} {name}"]
-            if status:
-                parts.append(status)
-            if fmt_short:
-                parts.append(f" : {fmt_short} ({paren_inner})" if paren_inner else f" : {fmt_short}")
-            elif paren_inner:
-                parts.append(f" ({paren_inner})")
-            lines.append("".join(parts))
-        return lines
-
-    def _get_nfo_files(self, meta: Meta) -> list[str]:
-        """Get NFO files in the release folder (including subdirectories).
-
-        Used by French trackers to include NFO files in .torrent and API upload."""
-        path = str(meta.get("path", ""))
-        if os.path.isdir(path):
-            # Directory release: search top-level first, then subdirectories (season packs)
-            nfo_files = glob.glob(os.path.join(path, "*.nfo"))
-            if not nfo_files:
-                nfo_files = glob.glob(os.path.join(path, "**", "*.nfo"), recursive=True)
-        else:
-            # Single-file release: only match an NFO with the same base name
-            stem = os.path.splitext(path)[0]
-            nfo_path = f"{stem}.nfo"
-            nfo_files = [nfo_path] if os.path.isfile(nfo_path) else []
-        if nfo_files:
-            meta["keep_nfo"] = True
-        return nfo_files
-
-    async def _patch_torrent_with_nfo(
-        self,
-        meta: Meta,
-        source_torrent_path: str,
-        nfo_files: list[str],
-    ) -> str | None:
-        """Create [tracker].torrent from an existing torrent + NFO without full rehash.
-
-        Appends NFO files to the END of the file list (after existing media
-        files) so that all piece hashes for the original content stay valid.
-        Only the last piece (which now includes NFO data) needs to be
-        recomputed by reading a few MB from disk instead of the full content.
-        """
-        import asyncio
-
-        from torf import Torrent
-
-        from src.console import console
-
-        try:
-            src = Torrent.read(source_torrent_path)
-            info = src.metainfo["info"]
-            piece_size: int = info["piece length"]
-            old_pieces_raw: bytes = info["pieces"]
-            old_files: list[dict[str, Any]] = info["files"]
-        except Exception:
-            return None
-
-        content_path = str(meta.get("path", ""))
-        if not os.path.isdir(content_path):
-            return None
-
-        old_piece_count = len(old_pieces_raw) // 20
-
-        tracker_name = getattr(self, "tracker", "")
-        source_flag = getattr(self, "source_flag", "")
-        tracker_config = self.config["TRACKERS"].get(tracker_name, {})  # type: ignore[attr-defined]
-        announce_url = str(tracker_config.get("announce_url", "https://fake.tracker")).strip()
-        if not tracker_name or not source_flag or not announce_url or announce_url == "https://fake.tracker":
-            return None
-        output_path = os.path.join(meta["base_dir"], "tmp", meta["uuid"], f"[{tracker_name}].torrent")
-
-        result = await asyncio.to_thread(
-            self._patch_torrent_with_nfo_sync,
-            src,
-            old_files,
-            old_pieces_raw,
-            old_piece_count,
-            piece_size,
-            content_path,
-            nfo_files,
-            source_flag,
-            announce_url,
-            output_path,
-        )
-        if result:
-            nfo_kb, tail_mb = result
-            console.print(f"[green]Patched torrent with NFO ({nfo_kb:.1f} KB) — read {tail_mb:.1f} MB instead of full rehash[/green]")
-            return output_path
-        return None
-
-    @staticmethod
-    def _patch_torrent_with_nfo_sync(
-        src: Any,
-        old_files: list[dict[str, Any]],
-        old_pieces_raw: bytes,
-        old_piece_count: int,
-        piece_size: int,
-        content_path: str,
-        nfo_files: list[str],
-        source_flag: str,
-        announce_url: str,
-        output_path: str,
-    ) -> tuple[float, float] | None:
-        """Synchronous core of _patch_torrent_with_nfo (runs in a thread)."""
-        from torf import Torrent
-
-        # Bail out if any NFO file already exists in the source torrent
-        existing_rel_paths = {tuple(f_info["path"]) for f_info in old_files if f_info.get("path")}
-
-        # Read NFO file data and build new file entries (appended at end)
-        nfo_entries: list[dict[str, Any]] = []
-        nfo_data = b""
-        for nfo_path in sorted(nfo_files):
-            # Compute path components relative to content_path
-            rel = os.path.relpath(nfo_path, content_path)
-            path_components = rel.replace("\\", "/").split("/")
-
-            # Skip NFOs that live outside the release tree: they have no valid
-            # in-torrent path. They are still delivered through the tracker's
-            # separate NFO upload field, exactly like the generated MediaInfo NFO.
-            if os.path.isabs(rel) or ".." in path_components:
-                continue
-
-            if tuple(path_components) in existing_rel_paths:
-                return None
-
-            try:
-                with open(nfo_path, "rb") as f:  # noqa: ASYNC230
-                    data = f.read()
-            except Exception:
-                return None
-            nfo_entries.append({"length": len(data), "path": path_components})
-            nfo_data += data
-
-        if not nfo_data:
-            return None
-
-        # Determine which piece is the last one containing existing file data
-        last_piece_idx = max(0, old_piece_count - 1)
-        last_piece_start = last_piece_idx * piece_size
-
-        # Read the data for the last piece from the content on disk.
-        # Walk through old files to find the file(s) that overlap with this piece.
-        offset = 0
-        last_piece_data = b""
-        for f_info in old_files:
-            f_length: int = f_info["length"]
-            f_name = os.path.join(content_path, *f_info["path"])
-            file_end = offset + f_length
-            if file_end <= last_piece_start:
-                offset = file_end
-                continue
-            # This file overlaps with the last piece
-            try:
-                actual_size = os.path.getsize(f_name)
-            except OSError:
-                return None
-            if actual_size != f_length:
-                return None
-            read_start = max(0, last_piece_start - offset)
-            try:
-                with open(f_name, "rb") as fh:  # noqa: ASYNC230
-                    fh.seek(read_start)
-                    last_piece_data += fh.read(f_length - read_start)
-            except Exception:
-                return None
-            offset = file_end
-
-        # Append all NFO data after the existing content
-        last_piece_data += nfo_data
-
-        # Compute piece hashes for this tail portion (usually just 1 piece)
-        new_tail_hashes = b""
-        for i in range(0, len(last_piece_data), piece_size):
-            chunk = last_piece_data[i : i + piece_size]
-            new_tail_hashes += hashlib.sha1(chunk, usedforsecurity=False).digest()  # nosec B324
-
-        # Build final pieces: keep unchanged prefix, replace tail
-        unchanged_prefix = old_pieces_raw[: last_piece_idx * 20]
-        final_pieces = unchanged_prefix + new_tail_hashes
-
-        # Build a copy of the source torrent with updated file list and pieces
-        patched = Torrent.copy(src)
-        patched.metainfo["info"]["files"] = list(old_files) + nfo_entries
-        patched.metainfo["info"]["pieces"] = final_pieces
-        patched.metainfo["info"]["source"] = source_flag
-        patched.metainfo["comment"] = ""
-        patched.metainfo["announce"] = announce_url
-
-        # Strip residual tracker/seed fields from the source torrent
-        for key in ("announce-list", "url-list", "httpseeds", "nodes"):
-            patched.metainfo.pop(key, None)
-
-        try:
-            patched.write(output_path, overwrite=True)
-        except Exception:
-            return None
-
-        nfo_kb = len(nfo_data) / 1024
-        tail_mb = len(last_piece_data) / (1024 * 1024)
-        return (nfo_kb, tail_mb)
-
-    @staticmethod
-    def _patch_mi_filename(mi_text: str, new_name: str) -> str:
-        """Replace the ‘Complete name’ value in MediaInfo text with *new_name*.
-
-        French trackers rename releases (language tags, notag label, French
-        title…), so the original filename inside a MediaInfo report no longer
-        matches the generated release name — some sites (e.g. C411) even
-        validate the two against each other.  This patches the ‘Complete
-        name’ line while preserving the file extension.
-        """
-        if not mi_text or not new_name:
-            return mi_text
-
-        def _replace_complete_name(match: re.Match[str]) -> str:
-            prefix = match.group(1)  # "Complete name    : "
-            old_value = match.group(2)
-            ext_match = re.search(r"(\.[a-zA-Z0-9]{2,4})$", old_value)
-            ext = ext_match.group(1) if ext_match else ""
-            return f"{prefix}{new_name}{ext}"
-
-        return re.sub(
-            r"^(Complete name\s*:\s*)(.+)$",
-            _replace_complete_name,
-            mi_text,
-            count=1,
-            flags=re.MULTILINE,
-        )
-
-    async def _get_mediainfo_text(self, meta: Meta) -> str:
-        """Read MediaInfo text from temp files (BDInfo/in-memory fallbacks).
-
-        The ‘Complete name’ line is patched to match the tracker-generated
-        release name (some sites, e.g. C411, check the two for consistency).
-        """
-        base = os.path.join(meta.get("base_dir", ""), "tmp", meta.get("uuid", ""))
-        content = ""
-
-        # Prefer clean-path, then standard mediainfo
-        for fname in ("MEDIAINFO_CLEANPATH.txt", "MEDIAINFO.txt"):
-            fpath = os.path.join(base, fname)
-            if os.path.exists(fpath):
-                async with aiofiles.open(fpath, encoding="utf-8") as f:
-                    content = await f.read()
-                    if content.strip():
-                        break
-                    content = ""
-
-        # BDInfo for disc releases
-        if not content and meta.get("bdinfo") is not None:
-            bd_path = os.path.join(base, "BD_SUMMARY_00.txt")
-            if os.path.exists(bd_path):
-                async with aiofiles.open(bd_path, encoding="utf-8") as f:
-                    return await f.read()
-
-        # Fallback: use in-memory mediainfo from prep
-        if not content:
-            fallback = str(meta.get("mediainfo_text") or "").strip()
-            if fallback:
-                content = fallback
-
-        if not content:
-            return ""
-
-        # Patch “Complete name” to match the tracker-generated release name
-        try:
-            name_result = await self.get_name(meta)
-            tracker_release_name = name_result.get("name", "") if isinstance(name_result, dict) else str(name_result)
-            if tracker_release_name:
-                content = self._patch_mi_filename(content, tracker_release_name)
-        except Exception:
-            pass  # If naming fails, return unpatched MI
-
-        return content
-
-    async def _get_source_description(self, meta: Meta) -> str:
-        """Reused/base description text (tmp DESCRIPTION.txt), already cleaned
-        by the description pipeline, for trackers that opt in via the
-        per-tracker ``include_source_description`` config flag.
-
-        The pipeline's own NFO embed ([spoiler=Scene NFO:] and the like) is
-        stripped: French trackers already send the NFO through their API
-        field, so repeating it in the description is pure redundancy.
-        """
-        if not self.config["TRACKERS"].get(self.tracker, {}).get("include_source_description", False):  # type: ignore[attr-defined]
-            return ""
-        path = os.path.join(str(meta.get("base_dir", "")), "tmp", str(meta.get("uuid", "")), "DESCRIPTION.txt")
-        try:
-            async with aiofiles.open(path, encoding="utf-8") as f:
-                content = await f.read()
-        except OSError:
-            return ""
-        content = re.sub(r"\[center\]\[spoiler=(?:Scene|FraMeSToR) NFO:\].*?\[/center\]", "", content, flags=re.DOTALL)
-        return content.strip()
-
-    async def _get_or_generate_nfo(self, meta: Meta) -> Union[str, None]:
-        """Pick the NFO to upload: the release's own NFO when present,
-        otherwise a MediaInfo file or a generated scene NFO.
-
-        Useful for trackers that expect an NFO with every upload (e.g. C411).
-        """
-        nfo_files = self._get_nfo_files(meta)
-        if nfo_files and not await is_multi_episode_nfo(nfo_files[0]):
-            return nfo_files[0]
-        else:
-            return await self._get_or_generate_mediainfo_as_nfo(meta)
-
-    async def _get_or_generate_mediainfo_as_nfo(self, meta: Meta) -> Union[str, None]:
-        """Sub-function of _get_or_generate_nfo to get MI file if exists
-        Else, generate a NFO
-        """
-        mi_dir = os.path.join(meta.get("base_dir", ""), "tmp", meta.get("uuid", ""))
-        mi_clean = os.path.join(mi_dir, "MEDIAINFO_CLEANPATH.txt")
-        mi = os.path.join(mi_dir, "MEDIAINFO.txt")
-        if os.path.isfile(mi_clean):
-            return mi_clean
-        elif os.path.isfile(mi):
-            return mi
-        else:
-            nfo_gen = SceneNfoGenerator(self.config)
-            return await nfo_gen.generate_nfo(meta, self.tracker)
-
-    @staticmethod
-    def _format_french_date(date_str: str) -> str:
-        """Format YYYY-MM-DD to French full date, e.g. 'jeudi 15 juillet 2010'."""
-        try:
-            dt = datetime.strptime(date_str, "%Y-%m-%d")
-            days = ["lundi", "mardi", "mercredi", "jeudi", "vendredi", "samedi", "dimanche"]
-            months = [
-                "",
-                "janvier",
-                "février",
-                "mars",
-                "avril",
-                "mai",
-                "juin",
-                "juillet",
-                "août",
-                "septembre",
-                "octobre",
-                "novembre",
-                "décembre",
-            ]
-            return f"{days[dt.weekday()]} {dt.day} {months[dt.month]} {dt.year}"
-        except (ValueError, IndexError):
-            return date_str
-
-    async def _recreated_torrent_if_nfo(self, meta: dict[str, Any], common: COMMON, config: dict[str, Any], tracker: str, source_flag: str) -> str:
-        """Re-create a .torrent if NFO is provided.
-
-        Some trackers requires the NFO if provided
-        by releaser. We generated a .torrent with
-        it if needed
-        """
-        nfo_files = self._get_nfo_files(meta)
-        if nfo_files:
-            upload_torrent_path = os.path.join(meta["base_dir"], "tmp", meta["uuid"], f"[{tracker}].torrent")
-
-            # Reuse existing torrent if it already contains .nfo files
-            if os.path.exists(upload_torrent_path):
-                try:
-                    from torf import Torrent
-
-                    existing = Torrent.read(upload_torrent_path)
-                    has_nfo = any(str(f).lower().endswith(".nfo") for f in existing.files)
-                    if has_nfo:
-                        meta["upload_torrent_path"] = upload_torrent_path
-                        return nfo_files[0]
-                except Exception:
-                    pass  # Fall through to recreation
-
-            # If BASE.torrent already contains NFO, clone it (no rehash needed)
-            base_torrent_path = os.path.join(meta["base_dir"], "tmp", meta["uuid"], "BASE.torrent")
-            if os.path.exists(base_torrent_path):
-                try:
-                    from torf import Torrent
-
-                    base = Torrent.read(base_torrent_path)
-                    if any(str(f).lower().endswith(".nfo") for f in base.files):
-                        await common.create_torrent_for_upload(meta, tracker, source_flag)
-                        meta["upload_torrent_path"] = upload_torrent_path
-                        return nfo_files[0]
-                except Exception:
-                    pass  # Fall through to full rehash
-
-            # Check if another tracker already created a torrent with NFO (avoid duplicate rehash)
-            tmp_dir = os.path.join(meta["base_dir"], "tmp", meta["uuid"])
-            for fname in os.listdir(tmp_dir):
-                if fname.startswith("[") and fname.endswith("].torrent") and fname != f"[{tracker}].torrent":
-                    try:
-                        from torf import Torrent
-
-                        other = Torrent.read(os.path.join(tmp_dir, fname))
-                        if any(str(f).lower().endswith(".nfo") for f in other.files):
-                            await common.create_torrent_for_upload(meta, tracker, source_flag, torrent_filename=fname.replace(".torrent", ""))
-                            meta["upload_torrent_path"] = upload_torrent_path
-                            return nfo_files[0]
-                    except Exception:  # nosec B112
-                        continue
-
-            # Patch an existing torrent by appending NFO to the file list.
-            # Only the last piece needs rehashing (a few MB instead of the full content).
-            patch_source = None
-            if os.path.exists(base_torrent_path):
-                patch_source = base_torrent_path
-            else:
-                # Try any tracker torrent as source
-                for fname in os.listdir(tmp_dir):
-                    if fname.startswith("[") and fname.endswith("].torrent") and fname != f"[{tracker}].torrent":
-                        patch_source = os.path.join(tmp_dir, fname)
-                        break
-            if patch_source:
-                try:
-                    patched = await self._patch_torrent_with_nfo(meta, patch_source, nfo_files)
-                    if patched and os.path.exists(patched):
-                        meta["upload_torrent_path"] = upload_torrent_path
-                        return nfo_files[0]
-                except Exception as e:
-                    console.print(f"[yellow]NFO patch failed ({e}), falling back to full rehash[/yellow]")
-
-            tracker_config = config["TRACKERS"].get(tracker, {})
-            tracker_url = str(tracker_config.get("announce_url", "https://fake.tracker")).strip()
-            torrent_create = f"[{tracker}]"
-            try:
-                cooldown = int(config.get("DEFAULT", {}).get("rehash_cooldown", 0) or 0)
-            except (ValueError, TypeError):
-                cooldown = 0
-            if cooldown > 0:
-                await asyncio.sleep(cooldown)
-            await TorrentCreator.create_torrent(meta, str(meta["path"]), torrent_create, tracker_url=tracker_url)
-            if not os.path.exists(upload_torrent_path):
-                raise FileNotFoundError(f"Failed to create {upload_torrent_path}")
-            meta["upload_torrent_path"] = upload_torrent_path
-            return nfo_files[0]
-        else:
-            return ""
