+++ src/trackers/french/__init__.py
@@ -1,0 +1,1 @@
+"""French tracker knowledge, split by concern (see each module's docstring for the upbrr counterpart)."""
